Deleting an organization adheres to business logic
Our original implementation of the endpoint `DELETE
/v3/organization/:guid` did not properly address business logic. For
example, when deleting an org it would blindly delete all private
domains owned by that org, regardless of whether they had been shared
with other orgs.

The new behavior checks first if a domain is shared. If it is, it
doesn't delete; instead, it returns an informative error.

In addition to being more circumspect with private domains, we address
the following issues:

- When the org to be deleted has private domains shared with it, those
domains are first unshared before deleting the organization.

- When service plans are associated with the organization, they are
first unassocated with the org before its deletion.

- Quota definitions owned by the org are deleted.

- Isolation segments entitled to the organization are unentitled.

- Users with roles in the organization are disassociated.

Developer Notes:

- We moved the OrganizationDelete action (and its associated spec) under
the `v2/` subdirectory in order to make room for new V3 action.

- We fixed the spelling of the word `inaccessible`, which had a possible
world record of three distinct typos. We also use `disassociate` rather
than `unassociate`.

[Finishes #163341376]

+++ app/actions/organization_delete.rb
@@ -16,8 +16,24 @@ module VCAP::CloudController
           return [CloudController::Errors::ApiError.new_from_details('OrganizationDeletionFailed', org.name, error_message)]
         end
 
+        domains_to_unshare = []
+        org.private_domains.each do |private_domain|
+          if private_domain.owning_organization == org
+            if private_domain.shared_with_any_orgs?
+              errs << "Domain '#{private_domain.name}' is shared with other organizations. Unshare before deleting."
+            end
+          else
+            domains_to_unshare << private_domain
+          end
+        end
+
+        unless errs.empty?
+          return [CloudController::Errors::ApiError.new_from_details('OrganizationDeletionFailed', org.name, errs.join("\n\n"))]
+        end
+
         Organization.db.transaction do
           delete_metadata(org)
+          unshare_private_domains(domains_to_unshare, org)
           org.destroy
         end
       end
@@ -33,6 +49,15 @@ module VCAP::CloudController
     def delete_metadata(org_model)
       LabelDelete.delete(org_model.labels)
       AnnotationDelete.delete(org_model.annotations)
+    end
+
+    def unshare_private_domains(domains_to_unshare, org_model)
+      # find all private domains that are shared with this org
+      # for each one, unshare the org
+
+      domains_to_unshare.each do |domain|
+        domain.remove_shared_organization(org_model)
+      end
     end
   end
 end
